+++ Goblet-of-Fire/src/cTransform.cpp
@@ -1,8 +1,8 @@
 #include "cTransform.hpp"
 
 namespace GobletOfFire::ObjectComponent {
-  cTransform::cTransform(uint32_t x, uint32_t y)
-    : point_(Physics::point2<uint32_t>(x, y)) {}
+  cTransform::cTransform(float x, float y)
+    : point_(Physics::point2<float>(x, y)) {}
 
   void cTransform::create() {}
   void cTransform::destroy() {}
@@ -12,15 +12,15 @@ namespace GobletOfFire::ObjectComponent {
   void cTransform::update(const Utilities::Time::duration) {}
   void cTransform::render(Graphics::buffer&) {}
 
-  void cTransform::setPosition(uint32_t x, uint32_t y) {
+  void cTransform::setPosition(float x, float y) {
     point_.x = x;
     point_.y = y;
   }
-  void cTransform::setPosition(Physics::point2<uint32_t> point) {
+  void cTransform::setPosition(Physics::point2<float> point) {
     point_ = point;
   }
 
-  Physics::point2<uint32_t> cTransform::getPoint() const {
+  Physics::point2<float> cTransform::getPoint() const {
     return point_;
   }
 }
